requirements.txt
bs4
methodtools
opencc-python-reimplemented
python-decouple
requests
requests_html
typer
selenium
